kata-deploy: Add drop-in conf support containerd
Upcoming containerd 2.0 will have drop-in conf support.
First patch to get this functionality into our stack.

+++ tools/packaging/kata-deploy/scripts/kata-deploy.sh
@@ -31,6 +31,7 @@ snapshotters_delimiter=':'
 
 AGENT_HTTPS_PROXY="${AGENT_HTTPS_PROXY:-}"
 AGENT_NO_PROXY="${AGENT_NO_PROXY:-}"
+CONTAINERD_DROP_IN_CONF="${CONTAINERD_DROP_IN_CONF:-}"
 
 # If we fail for any reason a message will be displayed
 die() {
@@ -365,7 +366,7 @@ function configure_containerd_runtime() {
 	
 	# if we are running k0s auto containerd.toml generation, the base template is by default version 2
 	# we can safely assume to reference the newer version of cri
-	if grep -q "version = 2\>" $containerd_conf_file || [ "$1" == "k0s-worker" ] || [ "$1" == "k0s-controller" ]; then
+	if grep -q "version = 2\>" $containerd_conf_file || [ "$1" == "k0s-worker" ] || [ "$1" == "k0s-controller" ] || [ -n "${CONTAINERD_DROP_IN_CONF}" ]; then
 		pluginid=\"io.containerd.grpc.v1.cri\"
 	fi
 	local runtime_table=".plugins.${pluginid}.containerd.runtimes.\"${runtime}\""
@@ -403,7 +404,8 @@ function configure_containerd() {
 
 	mkdir -p /etc/containerd/
 
-	if [ -f "$containerd_conf_file" ]; then
+	# If we have a drop-in configuration file, we don't need to touch the original
+	if [ -f "$containerd_conf_file" ] && [ -z "${CONTAINERD_DROP_IN_CONF}" ]; then
 		# backup the config.toml only if a backup doesn't already exist (don't override original)
 		cp -n "$containerd_conf_file" "$containerd_conf_file_backup"
 	fi
@@ -518,6 +520,7 @@ function main() {
 	echo "* SNAPSHOTTER_HANDLER_MAPPING: ${SNAPSHOTTER_HANDLER_MAPPING}"
 	echo "* AGENT_HTTPS_PROXY: ${AGENT_HTTPS_PROXY}"
 	echo "* AGENT_NO_PROXY: ${AGENT_NO_PROXY}"
+	echo "* CONTAINERD_DROP_IN_CONF: ${CONTAINERD_DROP_IN_CONF}"
 
 	# script requires that user is root
 	euid=$(id -u)
@@ -543,6 +546,8 @@ function main() {
 		# This works by k0s creating a special directory in /etc/k0s/containerd.d/ where user can drop-in partial containerd configuration snippets.
 		# k0s will automatically pick up these files and adds these in containerd configuration imports list.
 		containerd_conf_file="/etc/containerd/kata-containers.toml"
+	elif [ -n "${CONTAINERD_DROP_IN_CONF}" ]; then 
+		containerd_conf_file=/etc/containerd/"${CONTAINERD_DROP_IN_CONF}"
 	else
 		# runtime == containerd
 		if [ ! -f "$containerd_conf_file" ] && [ -d $(dirname "$containerd_conf_file") ] && \
